Modified the pcb.h file yet again!  Just a few minor tweaks, as well as a change to the Destroy method.
A few notes about the Destroy method: the method is now designed to free
the Priority and ID field for the PCB passed in, it then frees the PCB
itself, so the user doesn't have to worry about that.  Finally, it
returns the pointer to the Next PCB, it doesn't destroy it because it
seems like doing that would destroy the next PCB?  That isn't what we
want, so instead it gets returned.

I also pretty much finished the pcb.c file.  I'm sure it can be
improved, but I believe most of what we need is there.  Another note
though, I didn't write methods for the getter methods, because I
realized we don't need them, right?  We can just call pcb->field.  I'll
keep the method signatures in the .h file for now, but I don't think
they're necessary.

And that concludes my super long commit note!