Create priorityQueueTest.c
This is halfway done, only tests enqueueing.  There's something wonky with it though because whenever I try to create another PCB on the *second* iteration of the loop, I get a segmentation fault.  I haven't had time to figure out if that's a result of something in the newPCB method or my test code (probably in my test code!).