Update PriorityQueue.c
The toString wasn't quite working for my test, so I made some changes to it.  It's working for me now, but I'd make sure it works with your machines just in case!  Also slightly changed the constructor, I initialized FIFOs for each slot in the priorityArray (instead of them being initialized to null).  All the original code is there, in case something's wrong with my changes.