Finished test
The test is finished.  There's something going on with one of the toStrings, I'm thinking the PCB toString.  I can run this test just fine in my debugger, but when I try to run it normally, I get a crash.  I think I've narrowed it down to somewhere in the PCB toString (with the extensive use of printfs!), but I don't know what it is.  I think it's some kind of memory/stack breach, based on what I've read online.