Slight modification to pqDequeue
I just had to make a slight modification to the dequeue method.  This is only because of the change I made in the constructor, where a fifo queue is instantiated for each spot in the priority array.  This can be easily changed if you guys think that this isn't the way to do things.