Fixed the CPU.c crash
It was happening at line 116 (or the original code).  fifoQueuePeek was being called when the queue was empty, and the peek method was returning a NULL.  That NULL was then being passed into the PCBToString method, which doesn't know how to handle NULLs.  I think I'll go and fix that too actually, so that if we accidentally send a null into a toString method, we don't get a crash.