Update CPU.c
Fixed the problem with CPU running (where the ready queue always seems to be empty and the first created PCB is always the one running).  I just modified the dispatcher method to dequeue all the PCBs currently sitting in the newProcesses queue, and enqueue them into the readyProcesses queue.  I also just added some extra new lines into the printf statements in the main loop so that the output is a little easier to see (broken into chunks).