new main
Added in what I have so far for the main loop.  It relies on changes to the PCB class, which I did on my own, but if you haven't then it's going to throw errors.  Also I haven't run it yet, but I think the structure is correct so far.  Oh, as well as some methods that probably aren't in there yet, like io1Check() and io2Check().